requirements.txt
PyQt6>=6.4.0
pyfluidsynth>=1.3.2
anthropic>=0.28.0
openai>=1.30.0
pytest>=7.4.0
